+++ python/cuspatial/cuspatial/_lib/trajectory.pyx
@@ -1,9 +1,4 @@
-# Copyright (c) 2019, NVIDIA CORPORATION.
-
-# cython: profile=False
-# distutils: language = c++
-# cython: embedsignature = True
-# cython: language_level = 3
+# Copyright (c) 2019-2020, NVIDIA CORPORATION.
 
 from cudf._lib.cpp.column.column cimport column
 from cudf._lib.cpp.column.column_view cimport column_view
@@ -18,10 +13,9 @@ from cuspatial._lib.cpp.trajectory cimport (
     trajectory_distances_and_speeds as cpp_trajectory_distances_and_speeds,
 )
 
-from cuspatial._lib.move cimport move
-
 from libcpp.memory cimport unique_ptr
 from libcpp.pair cimport pair
+from libcpp.utility cimport move
 
 cpdef derive_trajectories(Column object_id, Column x,
                           Column y, Column timestamp):